히스토리 알고리즘 수정, 입장 거절 수정

+++ src/main/java/com/hawkeye/capstone/domain/History.java
@@ -37,7 +37,7 @@ public class History {
 
     private int attendanceCount;
 
-    private int attitude;
+    private float attitude;
 
     private int vibe;
